Add label buttons
This commit adds funtionality to display 10 label buttons and
a play button in the first column of the grid layout.
When a label button is clicked, both the label button and go button
change their text color and font weight.
However, further modifications are needed in the 'drawing annotation'
and 'add label' functions to appropriatley assign label name.