Fix label button and add label check function
Fixed label_button message box error.
Added frame_label_check function on dcm_data.py file.
Merged from main branch.
When returning to the frame with the label after moving the frame,
it needs to be modified so that the existing label is deleted
when the label button is pressed.